# Conflicts:
#	build.gradle
#	gradle/wrapper/gradle-wrapper.properties
#	src/main/java/az/gdg/msarticle/client/AuthenticationClient.java
#	src/main/java/az/gdg/msarticle/config/FeignConfig.java
#	src/main/java/az/gdg/msarticle/config/SecurityConfig.java
#	src/main/java/az/gdg/msarticle/config/SwaggerConfig.java
#	src/main/java/az/gdg/msarticle/filter/AuthenticationTokenFilter.java
#	src/main/java/az/gdg/msarticle/model/CommentRequest.java
#	src/main/java/az/gdg/msarticle/model/client/auth/HttpHeader.java
#	src/main/java/az/gdg/msarticle/model/client/auth/UserInfo.java
#	src/main/java/az/gdg/msarticle/model/dto/ArticleDTO.java
#	src/main/java/az/gdg/msarticle/model/dto/CommentDTO.java
#	src/main/java/az/gdg/msarticle/model/dto/TagDTO.java
#	src/main/java/az/gdg/msarticle/model/entity/ArticleEntity.java
#	src/main/java/az/gdg/msarticle/model/entity/CommentEntity.java
#	src/main/java/az/gdg/msarticle/model/entity/TagEntity.java
#	src/main/java/az/gdg/msarticle/repository/CommentRepository.java
#	src/main/java/az/gdg/msarticle/security/UserAuthentication.java
#	src/main/resources/application.yml

+++ src/main/java/az/gdg/msarticle/service/impl/ArticleServiceImpl.java
@@ -39,7 +39,7 @@ public class ArticleServiceImpl implements ArticleService {
         logger.info("ActionLog.getArticleById.start with id {}", articleId);
         ArticleEntity articleEntity = articleRepository.findById(articleId)
                 .orElseThrow(() -> new NoSuchArticleException("Article doesn't exist"));
-        Integer userId = articleEntity.getUserId();
+        Long userId = articleEntity.getUserId();
         if (articleEntity.isDraft()) {
             try {
                 if (Integer.parseInt(getAuthenticatedObject().getPrincipal().toString()) != userId) {
